# framework/deploy/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Builder Framework API
  version: 1.0.0
  description: |
    External REST API for the Knowledge Builder Framework.

    Two interaction models:

    **Consumption Flow** — `POST /api/v1/ask`
    Single endpoint for all knowledge queries. Routes internally through a four-tier
    confidence system (workflow skill → KB retrieval → multi-persona fanout → no-answer).
    Clients never specify which KB, retriever, or persona skill to use.

    **Knowledge Builder Flow** — `/api/v1/kb/authorSkill`
    Server-side stateful session API for authoring new skills. The client is a
    pass-through: it posts user input, shows the server's message, collects the next
    reply, and repeats until `done=true`. The server drives a 14-state machine.

    **Naming conventions:**
    - External API surface (this spec): camelCase field names.
    - DB columns: snake_case (PostgreSQL/Oracle convention). The API layer translates.
    - Python internal code: snake_case (Python convention). Not visible here.

    **Authentication:** Bearer token required on all endpoints.
    Token scopes: `read` (ask + GET authorSkill), `write` (POST/DELETE authorSkill), `admin`.

  contact:
    name: Knowledge Builder Framework Team
  license:
    name: Internal
servers:
  - url: https://{ociVmHostname}
    description: OCI Compute VM (SSE/HTTPS mode — typical for team members)
    variables:
      ociVmHostname:
        default: kb-framework.internal
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/ask:
    post:
      operationId: askKnowledgeBase
      tags: [Consumption]
      summary: Single entry point for all knowledge queries
      description: |
        Routes internally through the four-tier system:
        - Tier 1: Workflow skill match (≥0.85 confidence) — returns curated artifact output
        - Tier 2: KB retrieval + synthesis (≥0.60) — cited answer from raw content
        - Tier 3: Multi-persona fanout (≥0.40) — synthesized from multiple persona KBs
        - Tier 4: Honest "no answer" (<0.40) — includes skill suggestion

        `persona` and `serviceId` are routing hints, not hard filters.
        The intent classifier makes the final routing decision.

        MCP tool name: `askKnowledgeBase`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
            examples:
              opsIncidentQuery:
                summary: Query about recent incidents
                value:
                  question: "What were the root causes of the last 3 pod DB incidents?"
                  persona: "ops_eng"
                  serviceId: null
                  functionalArea: null
                  maxResults: 10
              tpmBlockerQuery:
                summary: TPM weekly blockers
                value:
                  question: "What were the top 3 blockers in the TPM weekly review last week?"
                  persona: "tpm"
      responses:
        '200':
          description: Synthesized answer with citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
              examples:
                tier1Match:
                  summary: Tier 1 workflow skill match
                  value:
                    answer: "Based on last week's exec review: 1. Customer Events migration blocked on DB schema approval. 2. Pod refresh behind schedule in EMEA. 3. Auth service dependency on Q2 deadline."
                    citations:
                      - contentId: "wiki-tpm-weekly-2026-05-05"
                        chunkId: "chunk-003"
                        text: "Blocker 1: Customer Events migration blocked on DB schema approval from infra team."
                        citationUrl: "https://confluence.internal/PROD/weekly-2026-05-05"
                        score: 0.91
                        metadata:
                          persona: "tpm"
                          kind: "wiki"
                          functionalArea: ["program_management"]
                          resources: []
                          services: ["customer-events"]
                          sourceSha: "a3f1b2c"
                          schemaVersion: 1
                    confidence: 0.91
                    tierUsed: 1
                    tierDescription: "workflow_skill"
                    costTokens:
                      prompt: 840
                      completion: 210
                      total: 1050
                    skillSuggestion:
                      offered: false
                      message: null
                      candidateId: null
                tier4NoAnswer:
                  summary: Tier 4 no answer with skill suggestion
                  value:
                    answer: "I couldn't find a grounded answer to this question."
                    citations: []
                    confidence: 0.28
                    tierUsed: 4
                    tierDescription: "no_answer"
                    costTokens:
                      prompt: 310
                      completion: 45
                      total: 355
                    skillSuggestion:
                      offered: true
                      message: "This query type has been logged. Want to scaffold a workflow skill for queries like this?"
                      candidateId: "cand-ops_eng-20260510-x7f2"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/UpstreamUnavailable'

  # ─────────────────────────────────────────────────────────────────────────
  # Knowledge Builder Flow — session management
  # ─────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token issued per consumer. Stored in OCI Vault.
        Format: `Authorization: Bearer <token>`

        Token scopes:
        - `read` — askKnowledgeBase and GET /api/v1/kb/authorSkill* endpoints
        - `write` — authorSkill (POST) and DELETE /api/v1/kb/authorSkill/{synthId}
        - `admin` — metrics and system configuration endpoints
  schemas:
...
    AskRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 4096
          description: The natural language question or task. Required.
          example: "What were the root causes of the last 3 pod DB incidents?"
        persona:
          type: [string, 'null']
          description: |
            Routing hint — restrict to this persona's shim scope.
            The intent classifier may override this if it contradicts a high-confidence
            classification. Hint only, not a hard filter.
          example: "ops_eng"
        serviceId:
          type: [string, 'null']
          description: Routing hint — restrict to content tagged with this service.
          example: "customer-events"
        functionalArea:
          type: [string, 'null']
          description: Routing hint — restrict to this functional area dimension.
          example: "database_operations"
        maxResults:
          type: [integer, 'null']
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum KB results to gather per persona.
    AskResponse:
      type: object
      required: [answer, citations, confidence, tierUsed, tierDescription, costTokens, skillSuggestion]
      properties:
        answer:
          type: string
          description: Synthesized natural language answer with inline citations.
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
          description: Source evidence passages used to synthesize the answer.
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Overall confidence of the answer.
        tierUsed:
          type: integer
          enum: [1, 2, 3, 4]
          description: |
            Which tier produced the answer:
            - 1: workflow skill match
            - 2: KB retrieval
            - 3: multi-persona fanout
            - 4: no answer
        tierDescription:
          type: string
          enum: [workflow_skill, kb_retrieval, multi_persona_fanout, no_answer]
          description: Human-readable tier name.
        costTokens:
          $ref: '#/components/schemas/TokenCost'
        skillSuggestion:
          $ref: '#/components/schemas/SkillSuggestion'
    Citation:
      type: object
      required: [contentId, text, citationUrl, score, metadata]
      properties:
        contentId:
          type: string
          description: Unique ID of the source ContentItem.
        chunkId:
          type: [string, 'null']
          description: ID of the specific chunk within the ContentItem, if applicable.
        text:
          type: string
          description: The evidence passage from the source.
        citationUrl:
          type: string
          format: uri
          description: |
            Source URL. Always present — absence is a bug.
            Points to the original Confluence page, Jira ticket, git file, etc.
        score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Retrieval confidence score for this citation.
        metadata:
          $ref: '#/components/schemas/CitationMetadata'
    CitationMetadata:
      type: object
      required: [persona, kind, sourceSha, schemaVersion]
      properties:
        persona:
          type: string
          description: Persona KB this content came from.
        kind:
          type: string
          enum: [vector, wiki, graph, sql_passthrough, code_index, filestore]
          description: Storage kind of the source KB.
        functionalArea:
          type: array
          items:
            type: string
          description: Functional area tags on this content.
        resources:
          type: array
          items:
            type: string
          description: Resource tags on this content.
        services:
          type: array
          items:
            type: string
          description: Service tags on this content.
        sourceSha:
          type: string
          description: Git SHA or content hash of the source at ingestion time.
        schemaVersion:
          type: integer
          description: Parser schema version used to extract this content.
    TokenCost:
      type: object
      required: [prompt, completion, total]
      properties:
        prompt:
          type: integer
          minimum: 0
          description: Prompt tokens consumed.
        completion:
          type: integer
          minimum: 0
          description: Completion tokens consumed.
        total:
          type: integer
          minimum: 0
          description: Total tokens consumed.
    SkillSuggestion:
      type: object
      required: [offered]
      properties:
        offered:
          type: boolean
          description: True if Tier 4 triggered a skill suggestion.
        message:
          type: [string, 'null']
          description: Suggestion message to show the user.
        candidateId:
          type: [string, 'null']
          description: ID of the logged skill candidate in kb_shim.skill_candidates.

    # ── AuthorSkill — requests ───────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code.
              enum:
                - invalid_argument
                - invalid_input
                - unauthenticated
                - permission_denied
                - session_not_found
                - session_expired
                - session_done
                - rate_limited
                - upstream_unavailable
                - budget_exceeded
            message:
              type: string
              description: Human-readable error message.
            details:
              type: object
              description: Optional additional error context.
              additionalProperties: true
  responses:
    BadRequest:
      description: |
        Bad request. Common causes:
        - question is empty or exceeds 4096 chars (ask endpoint)
        - input does not match accepted options for current state (authorSkill)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "invalid_argument"
              message: "question must be between 1 and 4096 characters"
              details: {}
    Unauthenticated:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthenticated"
              message: "Authorization header missing or token invalid"
              details: {}
    PermissionDenied:
      description: Token lacks the required scope for this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "permission_denied"
              message: "Token lacks 'write' scope required for this operation"
              details: {}
    RateLimited:
      description: Per-token RPM cap exceeded.
      headers:
        Retry-After:
          description: Seconds until the rate limit window resets.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "rate_limited"
              message: "Rate limit exceeded. Retry after 23 seconds."
              details:
                retryAfterSeconds: 23
    UpstreamUnavailable:
      description: ADB or OCI GenAI is unreachable. Retry with exponential backoff.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "upstream_unavailable"
              message: "Oracle ADB is currently unreachable. The framework is in read-only mode."
              details:
                affectedComponent: "adb"